improve live query subscription files
