+++ Sources/ParseSwift/LiveQuery/Subscription.swift
@@ -7,60 +7,13 @@
 //
 //
 
+#if canImport(Combine)
 import Foundation
 
 /**
- Represents an update on a specific object from the `ParseLiveQuery` Server.
- - Entered: The object has been updated, and is now included in the query.
- - Left:    The object has been updated, and is no longer included in the query.
- - Created: The object has been created, and is a part of the query.
- - Updated: The object has been updated, and is still a part of the query.
- - Deleted: The object has been deleted, and is no longer included in the query.
- */
-public enum Event<T: ParseObject> {
-    /// The object has been updated, and is now included in the query.
-    case entered(T)
-
-    /// The object has been updated, and is no longer included in the query.
-    case left(T)
-
-    /// The object has been created, and is a part of the query.
-    case created(T)
-
-    /// The object has been updated, and is still a part of the query.
-    case updated(T)
-
-    /// The object has been deleted, and is no longer included in the query.
-    case deleted(T)
-
-    init?(event: EventResponse<T>) {
-        switch event.op {
-        case .enter: self = .entered(event.object)
-        case .leave: self = .left(event.object)
-        case .create: self = .created(event.object)
-        case .update: self = .updated(event.object)
-        case .delete: self = .deleted(event.object)
-        default: fatalError()
-        }
-    }
-}
-
-private func == <T>(lhs: Event<T>, rhs: Event<T>) -> Bool {
-    switch (lhs, rhs) {
-    case (.entered(let obj1), .entered(let obj2)): return obj1 == obj2
-    case (.left(let obj1), .left(let obj2)):       return obj1 == obj2
-    case (.created(let obj1), .created(let obj2)): return obj1 == obj2
-    case (.updated(let obj1), .updated(let obj2)): return obj1 == obj2
-    case (.deleted(let obj1), .deleted(let obj2)): return obj1 == obj2
-    default: return false
-    }
-}
-
-#if canImport(Combine)
-/**
  A default implementation of the `QuerySubscribable` protocol. Suitable for `ObjectObserved`
  as the subscription can be used as a SwiftUI publisher. Meaning it can serve
- indepedently as a ViewModel in MVVM.
+ indepedently as a ViewModel in MVVM. Also can be used as a Combine publisher.
  */
 open class Subscription<T: ParseObject>: QueryViewModel<T>, QuerySubscribable {
 
@@ -112,7 +65,11 @@ open class Subscription<T: ParseObject>: QueryViewModel<T>, QuerySubscribable {
         self.event = nil
         self.unsubscribed = nil
     }
+}
 
+// MARK: QuerySubscribable
+
+extension Subscription {
     open func didReceive(_ eventData: Data) throws {
         // Need to decode the event with respect to the `ParseObject`.
         let eventMessage = try ParseCoding.jsonDecoder().decode(EventResponse<T>.self, from: eventData)
@@ -131,30 +88,3 @@ open class Subscription<T: ParseObject>: QueryViewModel<T>, QuerySubscribable {
     }
 }
 #endif
-
-extension SubscriptionCallback {
-
-    /**
-     Register a callback for when an event occurs of a specific type
-     Example:
-         subscription.handle(Event.Created) { query, object in
-            // Called whenever an object is creaated
-         }
-     - parameter eventType: The event type to handle. You should pass one of the enum cases in `Event`.
-     - parameter handler: The callback to register.
-     - returns: The same subscription, for easy chaining.
-     */
-    @discardableResult public func handle(_ eventType: @escaping (T) -> Event<T>,
-                                          _ handler: @escaping (Query<T>, T) -> Void) -> SubscriptionCallback {
-        return handleEvent { query, event in
-            switch event {
-            case .entered(let obj) where eventType(obj) == event: handler(query, obj)
-            case .left(let obj)  where eventType(obj) == event: handler(query, obj)
-            case .created(let obj) where eventType(obj) == event: handler(query, obj)
-            case .updated(let obj) where eventType(obj) == event: handler(query, obj)
-            case .deleted(let obj) where eventType(obj) == event: handler(query, obj)
-            default: return
-            }
-        }
-    }
-}
