--- Sources/ParseSwift/LiveQuery/SubscriptionPublisher.swift
@@ -1,14 +1,0 @@
-//
-//  SubscriptionPublisher.swift
-//  ParseSwift
-//
-//  Created by Corey Baker on 11/3/21.
-//  Copyright © 2021 Parse Community. All rights reserved.
-//
-
-
-#if canImport(Combine)
-import Foundation
-import Combine
-
-#endif
